moved splash_screen to new cpp
added header files
